products: documentation with swagger

+++ src/modules/products/useCases/deleteProducts/DeleteProductsUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from "tsyringe";
 
 import { IStorageProvider } from "../../../../shared/container/providers/StorageProvider/IStorageProvider";
+import { AppError } from "../../../../shared/errors/AppError";
 import { IProductsRepository } from "../../repositories/IProductsRepository";
 
 @injectable()
@@ -13,6 +14,12 @@ class DeleteProductUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const productExists = await this.productsRepository.findById(id);
+
+    if (!productExists) {
+      throw new AppError("Product does not exists!");
+    }
+
     const product = await this.productsRepository.findById(id);
 
     await this.storageProvider.delete(product.image, "products");
